added negative scenario
