+++ integration_tests/helpers/commonHelpers.ts
@@ -2,21 +2,18 @@ import {
   ClientId,
   SNSEventPayload,
   EventName,
-  prettyClientNameMap,
-  prettyEventNameMap,
-  
 } from "../payloads/snsEventPayload";
 import { deleteObjectInS3, getS3ItemsList } from "./s3Helper";
 import { publishSNS } from "./snsHelper";
 import { resourcePrefix } from "../helpers/envHelper";
-import { startQueryExecutionCommand, queryObject } from "./athenaHelper";
+
 
 
 
 
 const prefix = resourcePrefix();
 const objectsPrefix = "btm_transactions";
-const databaseName = `${prefix}-calculations`;
+
 
 
 export const generateRandomId = (): string => {
@@ -145,26 +142,5 @@ export const deleteS3Events = async (
   return true;
 };
 
-export const queryResults = async({
-  clientId,
-  eventName, tableName
-}: {
-  clientId: ClientId;
-  eventName: EventName;
-  tableName:TableNames
-}): Promise<[]> => {
-  const prettyClientName = prettyClientNameMap[clientId];
-  const prettyEventName = prettyEventNameMap[eventName];
 
-  // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
-  const curatedQueryString = `SELECT * FROM "${tableName}" WHERE vendor_name='${prettyClientName}' AND service_name='${prettyEventName}'`;
-  console.log("🚀 ~ file: billing-transaction-view-tests.ts:101 ~ curatedQueryString", curatedQueryString)
-  const queryId = await startQueryExecutionCommand(
-    
-    databaseName,
-    curatedQueryString
-  );
-  const results = await queryObject(queryId);
-  return results;
-}
