+++ integration_tests/payloads/receiptNOValue.txt
@@ -1,0 +1,1 @@
+{"invoice_receipt_id":"receipt_0000","vendor_name":"Client two","total":6.6600,"invoice_receipt_date":"2023-01-06","subtotal":0.0000,"item_id":0,"item_description":"item_type_0000","service_name":"Passport check","unit_price":3.33,"quantity":0,"price":0.0000}
